0.1.5

+++ libNoRSX/Image.cpp
@@ -1,4 +1,4 @@
-#include "Image.h"
+#include <NoRSX/Image.h>
 
 void Image::LoadPNG(const char* filename, pngData *png){
 	pngLoadFromFile(filename, png);
@@ -20,20 +20,20 @@ void Image::DrawIMG(int x, int y, pngData *png1){
 
 	if(png1->bmp_out){
 		u32 *scr = (u32 *)G->buffers[G->currentBuffer].ptr;
-		u32 *png= (u32 *)(void *)png1->bmp_out;
+		u32 *png = (u32 *)(void *)png1->bmp_out;
 		unsigned int n, m;
 
 		scr += y*G->buffers[G->currentBuffer].width+x;
 		u32 height = png1->height;
-		if((y+height) >= G->buffers[G->currentBuffer].height)
+		if((y+height) >= (u32)G->buffers[G->currentBuffer].height)
 			height = G->buffers[G->currentBuffer].height;
 
 		u32 width = png1->width;
-		if((x+width) >=G->buffers[G->currentBuffer].width)
+		if((x+width) >= (u32)G->buffers[G->currentBuffer].width)
 			width = G->buffers[G->currentBuffer].width;
 
-		for(n=0;n < height;n++){
-			for(m=0;m < width;m++)
+		for(n=0;n < (u32)height;n++){
+			for(m=0;m < (u32)width;m++)
 				scr[m]=png[m];
 			png+=png1->pitch>>2;
 			scr+=G->buffers[G->currentBuffer].width;
@@ -44,20 +44,20 @@ void Image::DrawIMG(int x, int y, pngData *png1){
 void Image::DrawIMG(int x, int y, jpgData *jpg1){
 	if(jpg1->bmp_out){
 		u32 *scr = (u32 *)G->buffers[G->currentBuffer].ptr;
-		u32 *jpg= (u32 *)(void *)jpg1->bmp_out;
+		u32 *jpg = (u32 *)(void *)jpg1->bmp_out;
 		unsigned int n, m;
 
 		scr += y*G->buffers[G->currentBuffer].width+x;
 		u32 height = jpg1->height;
-		if((y+height) >= G->buffers[G->currentBuffer].height)
+		if((y+height) >= (u32)G->buffers[G->currentBuffer].height)
 			height = G->buffers[G->currentBuffer].height;
 
 		u32 width = jpg1->width;
-		if((x+width) >=G->buffers[G->currentBuffer].width)
+		if((x+width) >= (u32)G->buffers[G->currentBuffer].width)
 			width = G->buffers[G->currentBuffer].width;
 
-		for(n=0;n < height;n++){
-			for(m=0;m < width;m++)
+		for(n=0;n < (u32)height;n++){
+			for(m=0;m < (u32)width;m++)
 				scr[m]=jpg[m];
 			jpg+=jpg1->pitch>>2;
 			scr+=G->buffers[G->currentBuffer].width;
@@ -68,20 +68,20 @@ void Image::DrawIMG(int x, int y, jpgData *jpg1){
 void Image::AlphaDrawIMG(int x, int y, pngData *png1){
 	if(png1->bmp_out){
 		u32 *scr = (u32 *)G->buffers[G->currentBuffer].ptr;
-		u32 *png= (u32 *)(void *)png1->bmp_out;
+		u32 *png = (u32 *)(void *)png1->bmp_out;
 		unsigned int n, m;
 
 		scr += y*G->buffers[G->currentBuffer].width+x;
 		u32 height = png1->height;
-		if((y+height) >= G->buffers[G->currentBuffer].height)
+		if((y+height) >= (u32)G->buffers[G->currentBuffer].height)
 			height = G->buffers[G->currentBuffer].height;
 
 		u32 width = png1->width;
-		if((x+width) >=G->buffers[G->currentBuffer].width)
+		if((x+width) >= (u32)G->buffers[G->currentBuffer].width)
 			width = G->buffers[G->currentBuffer].width;
 
-		for(n=0;n < height;n++){
-			for(m=0;m < width;m++){
+		for(n=0;n < (u32)height;n++){
+			for(m=0;m < (u32)width;m++){
 				unsigned int a = png[m] >> 24;	 // alpha 
 				u32 OxFF_A = 0xff - a;
 				if (0 != a) 
@@ -115,15 +115,15 @@ void Image::AlphaDrawIMG(int x, int y, pngData *png1){
 void Image::DrawPartialImage(int x, int y, unsigned int s_width, unsigned int s_height, unsigned int e_width, unsigned int e_height, unsigned int bg, unsigned int color, pngData *png1){
 	if(png1->bmp_out){
 		u32 *scr = (u32 *)G->buffers[G->currentBuffer].ptr;
-		u32 *png= (u32 *)(void *)png1->bmp_out;
+		u32 *png = (u32 *)(void *)png1->bmp_out;
 		unsigned int n, m;
 
 		scr += y*G->buffers[G->currentBuffer].width+x;
 		if(s_height>0) for(unsigned int i=0;i<s_height;i++) png+=png1->pitch>>2;
 		for(n=s_height;n<e_height+s_height;n++){
-			if((y+n)>=G->buffers[G->currentBuffer].height) break;
+			if((y+n)>=(u32)G->buffers[G->currentBuffer].height) break;
 			for(m=s_width;m<e_width+s_width;m++){
-				if((x+m)>=G->buffers[G->currentBuffer].width) break;
+				if((x+m)>=(u32)G->buffers[G->currentBuffer].width) break;
 				if(png[m]!=bg) scr[m-s_width]=color;
 			}
 			png+=png1->pitch>>2;
@@ -135,15 +135,15 @@ void Image::DrawPartialImage(int x, int y, unsigned int s_width, unsigned int s_
 void Image::DrawPartialImage(int x, int y, unsigned int s_width, unsigned int s_height, unsigned int e_width, unsigned int e_height, unsigned int bg, unsigned int color, jpgData *jpg1){
 	if(jpg1->bmp_out){
 		u32 *scr = (u32 *)G->buffers[G->currentBuffer].ptr;
-		u32 *jpg= (u32 *)(void *)jpg1->bmp_out;
+		u32 *jpg = (u32 *)(void *)jpg1->bmp_out;
 		unsigned int n, m;
 
 		scr += y*G->buffers[G->currentBuffer].width+x;
 		if(s_height>0) for(unsigned int i=0;i<s_height;i++) jpg+=jpg1->pitch>>2;
 		for(n=s_height;n<e_height+s_height;n++){
-			if((y+n)>=G->buffers[G->currentBuffer].height) break;
+			if((y+n)>=(u32)G->buffers[G->currentBuffer].height) break;
 			for(m=s_width;m<e_width+s_width;m++){
-				if((x+m)>=G->buffers[G->currentBuffer].width) break;
+				if((x+m)>=(u32)G->buffers[G->currentBuffer].width) break;
 				if(jpg[m]!=bg) scr[m-s_width]=color;
 			}
 			jpg+=jpg1->pitch>>2;
@@ -155,20 +155,20 @@ void Image::DrawPartialImage(int x, int y, unsigned int s_width, unsigned int s_
 void Image::DrawIMGtoBitmap(int x, int y, pngData *png1, NoRSX_Bitmap *a){
 	if(png1->bmp_out){
 		u32 *scr = (u32 *)a->bitmap;
-		u32 *png= (u32 *)(void *)png1->bmp_out;
+		u32 *png = (u32 *)(void *)png1->bmp_out;
 		unsigned int n, m;
 
 		scr += y*G->buffers[G->currentBuffer].width+x;
 		u32 height = png1->height;
-		if((y+height) >= G->buffers[G->currentBuffer].height)
+		if((y+height) >= (u32)G->buffers[G->currentBuffer].height)
 			height = G->buffers[G->currentBuffer].height;
 
 		u32 width = png1->width;
-		if((x+width) >=G->buffers[G->currentBuffer].width)
+		if((x+width) >= (u32)G->buffers[G->currentBuffer].width)
 			width = G->buffers[G->currentBuffer].width;
 
-		for(n=0;n < height;n++){
-			for(m=0;m < width;m++)
+		for(n=0;n < (u32)height;n++){
+			for(m=0;m < (u32)width;m++)
 				scr[m]=png[m];
 			png+=png1->pitch>>2;
 			scr+=G->buffers[G->currentBuffer].width;
@@ -180,20 +180,20 @@ void Image::DrawIMGtoBitmap(int x, int y, pngData *png1, NoRSX_Bitmap *a){
 void Image::DrawIMGtoBitmap(int x, int y, jpgData *jpg1, NoRSX_Bitmap *a){
 	if(jpg1->bmp_out){
 		u32 *scr = (u32 *)a->bitmap;
-		u32 *jpg= (u32 *)(void *)jpg1->bmp_out;
+		u32 *jpg = (u32 *)(void *)jpg1->bmp_out;
 		unsigned int n, m;
 
 		scr += y*G->buffers[G->currentBuffer].width+x;
 		u32 height = jpg1->height;
-		if((y+height) >= G->buffers[G->currentBuffer].height)
+		if((y+height) >= (u32)G->buffers[G->currentBuffer].height)
 			height = G->buffers[G->currentBuffer].height;
 
 		u32 width = jpg1->width;
-		if((x+width) >=G->buffers[G->currentBuffer].width)
+		if((x+width) >= (u32)G->buffers[G->currentBuffer].width)
 			width = G->buffers[G->currentBuffer].width;
 
-		for(n=0;n < height;n++){
-			for(m=0;m < width;m++)
+		for(n=0;n < (u32)height;n++){
+			for(m=0;m < (u32)width;m++)
 				scr[m]=jpg[m];
 			jpg+=jpg1->pitch>>2;
 			scr+=G->buffers[G->currentBuffer].width;
@@ -210,15 +210,15 @@ void Image::AlphaDrawIMGtoBitmap(int x, int y, pngData *png1, NoRSX_Bitmap *a){
 
 		scr += y*G->buffers[G->currentBuffer].width+x;
 		u32 height = png1->height;
-		if((y+height) >= G->buffers[G->currentBuffer].height)
+		if((y+height) >= (u32)G->buffers[G->currentBuffer].height)
 			height = G->buffers[G->currentBuffer].height;
 
 		u32 width = png1->width;
-		if((x+width) >=G->buffers[G->currentBuffer].width)
+		if((x+width) >= (u32)G->buffers[G->currentBuffer].width)
 			width = G->buffers[G->currentBuffer].width;
 
-		for(n=0;n < height;n++){
-			for(m=0;m < width;m++){
+		for(n=0;n < (u32)height;n++){
+			for(m=0;m < (u32)width;m++){
 				unsigned int a = png[m] >> 24;	 // alpha 
 				u32 OxFF_A = 0xff - a;
 				if (0 != a)
